+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 5.5.0
+## Updates
+* Swift 5
+* Set minimum required version from ios 9.0 to 9.1
+* Replaced photo picker framework (Removed DKImagePickerController)
+* Replaced photo viewer (Removed Lightbox)
+
 # 5.4.0
 ## Updates
 * Updated default launcher icon and rebranding